requirements.txt
opencv-python
opencv-python-headless
dlib
face_recognition
obs-websocket-py
pillow
imgui[glfw]
numpy
imutils
flask
flask_socketio
